frontend/requirements.txt
Flask==2.2.2
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
Pillow==9.0.0
numpy==1.24.0
opencv-python==4.6.0.66

